// public-ui/src/publicGenerated/model/conceptAnalysis.ts
/**
 * AllOfUs Public API
 * The API for the AllOfUs data browser and public storefront.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Analysis } from './analysis';


export interface ConceptAnalysis {
    /**
     * concept id
     */
    conceptId: string;

    /**
     * gender analysis
     */
    genderAnalysis?: Analysis;

    /**
     * age analysis
     */
    ageAnalysis?: Analysis;

}